Método listarMeusEventos movido para MeusEventosAdapter
